+++ csharp/ToolGood.Algorithm2/AlgorithmEngineEx.cs
@@ -521,7 +521,7 @@ namespace ToolGood.Algorithm
                     LastError = obj.ErrorMsg;
                     return def;
                 }
-                return (ulong)obj.IntValue;
+                return (ulong)obj.LongValue;
             } catch (Exception ex) {
                 LastError = ex.Message + "\r\n" + ex.StackTrace;
             }
@@ -584,7 +584,7 @@ namespace ToolGood.Algorithm
                     LastError = obj.ErrorMsg;
                     return def;
                 }
-                return obj.IntValue;
+                return obj.LongValue;
             } catch (Exception ex) {
                 LastError = ex.Message + "\r\n" + ex.StackTrace;
             }
@@ -822,7 +822,7 @@ namespace ToolGood.Algorithm
                     LastError = obj.ErrorMsg;
                     return def;
                 }
-                return (ulong)obj.IntValue;
+                return (ulong)obj.LongValue;
             } catch (Exception ex) {
                 LastError = ex.Message + "\r\n" + ex.StackTrace;
             }
@@ -885,7 +885,7 @@ namespace ToolGood.Algorithm
                     LastError = obj.ErrorMsg;
                     return def;
                 }
-                return obj.IntValue;
+                return obj.LongValue;
             } catch (Exception ex) {
                 LastError = ex.Message + "\r\n" + ex.StackTrace;
             }
